Changed so that alignment of the filenames in tabs isn't hardwired.  I edit lots of files that have the same prefix, e.g. mmc_ext_XXX.ml, and if too many tabs get opened then I end up with 25 identical tabs.  This is much less of a problem if the text is right-aligned, since the .ml suffixes take up less space.
You can try this out by setting the resource "*tab.alignment: XmALIGNMENT_END".

One improvement would be to also put the "*" that indicates a file is modified
on the right if the titles are right-aligned.  I tried to implement that but for
some reason couldn't read the setting of the alignment resource successfully.