I was asked to elaborate on this commit. So here goes:
This commit is a mistake because the codebase is made worse by it. The code
removed is of higher quality than the average commit, is of use to many
users of NEdit and would have been the basis for even more improvements in
code quality and usability.

Ironically, this commit is the result of all the things it claimed to heal:
That one developer considers the repository his personal playground and
forces his view onto others. While there is no basis for the original
complaint, the aggressive tone, stubborn insistence and insulting nature of
the comeback makes it clear that this something much different than honest
care about the community process. What could have been a minor nuisance was
blown out of all proportation and made to a personal vendetta.

At the same time, any attempts to get to a closure were not ignored, but
actively impeded, any calming voice was shouted down.

Despite all this, I figured the commit would put an end to the unpleasant
situation and would offer the opportunity to restart. Despite the attempts
to prevent this, I'd still rather work on NEdit than defend against personal
attacks.

So with this commit I take back the added features Hidden Files (SF#502840),
Wrap Hairline (SF#701337 variant) and Smart Case Search (SF#685618)
committed on 2004-10-01.