Fix for SF#464967.
The area redisplayed from within TextDRedisplayRange() is appended by one
character. This is necessary to let the logic in redisplayLine() know that
the caret should be redrawn (hasCursor = true) which in turn triggers the
redraw by calling drawCursor() later in the function.

Both representations of the bug described in the bug report plus two more
for multi-line redraws are fixed. I noticed no drawbacks.