Fix for SF#1368576
The fix for SF#464967 increased the area processed by redisplayLine() by
one to let it know that the cursor should be redrawn. This was done by
increasing the endIndex given to redisplayLine() by one. Sadly, endIndex is
INT_MAX in some cases, which prevented redisplays completely in certain
cases. This fix moves the incrementation to a saner place.

Neither the bugs fixed with SF#464967 nor any of the examples given in
SF#1368576's description are visible. Again, I notice no drawbacks.