Fix for SF#1381893
As TK pointed out in the original bug report (SF#557032), internal state is
not adequate to keep the file history consistent if more than one session
is running. Once you accept that external state must be checked you don't
need to keep internal state at all.

So this patch removes the prevOpenMenuValid flag from WindowInfo and the
rest of the code. Validity of the menu is determined using only the
directory entry of the history file.