Fix multiple memory leaks in interpret.c:PromoteToGlobal()
InstallSymbol() allocates a new symbol and put this to the GlobalSymList, the
symbol passed to PromoteToGlobal() is lost.

This adds the symbol directly to the GlobalSymList, without calling
InstallSymbol().

We further prints a warning to the console for what I assume are parsing
errors, in case the symbol is in already the GlobalSymList.

Closes SF#1894726