remove argument limit from action routines
In the old days function calls could only have 9 arguments. This limitation
comes from the macro language. This limitation was removed but the code for
calling a action routine was not adopted to this change. So a buffer overflow
was created. This commit removes this buffer overflow by creating the buffer
dynamically from the number of arguments.

Closes SF BUG#2074318