EWMH raise support
From: Markus Schwarzenberg <schwarzb@ipms.fraunhofer.de>

This patch sets an EWMH hint before XMapRaised is called to raise a window
from within nedit (or nedit-server).

This is usefull for nedit in non-tabbed mode or nc when trying to
access/raise an already opened, but (partially) hidden nedit window,
especially when the "focus stealing prevention" of the window manager
cannot be configured to allow nedit to raise it's windows.
(Or more drastically: it makes nc usable again for those window managers)

The code for this patch is borrowed from wmctrl (1.0.7).

The patch is tested successfully with Fedora 9 / Metacity (Gnome).

SourceForge-patch-Id: 2337783