Fix font selector rejecting fonts with multiple copies
XListFonts() may return "actual" font count greater than "maxnames"
if the system has more than one copy of the selected font.
In fontsel.c function okAction [and startupFont]
this was causing the test for actual==1 to fail when clicking OK
for such a font.  Also the test was not checking for XListFonts() returning
pointer==NULL (which is the primary indication of failure, not actual==0).