Performance fixes for large (500MB+) tag files
Don't attempt to normalize relative pathnames on the fly when looking for
duplicate tags, this is causing us to go crazy and makes large tags files
take a few minutes to load.

Increase the default hash bucket size and change the tag structure to
waste less space on 64-bit systems.